slide-bar, wordpress testing & equities/currency pie charts
added a slide-bar & wordpress file. Couldn’t get the javascript to link
to wordpress so gave up. files still there though. Slide-bar added on
the right of the main.html page for the risk bar.
a section in the portfolio section was added for equities and currency,
and two pie charts above these tables was added.