more about section & correcting CSS errors
added more content to the about section. In doing so the CSS required
updating to stop things like the H2’s being formatted to a narrow box.
Corrected, along with the colour of the text, so the whole document
text was not all appearing white and each individual element needing
text colour formatting.